feat: #15 - 스터디 모집 & 스터디 그룹 기능 개발

+++ src/main/java/com/threego/algo/studyrecruit/query/dto/StudyRecruitCommentDTO.java
@@ -8,9 +8,9 @@ import lombok.*;
 @AllArgsConstructor
 public class StudyRecruitCommentDTO {
 
-    private Long id;                    // 댓글 ID
-    private Long postId;                // 게시글 ID
-    private Long parentId;              // 부모 댓글 ID (대댓글인 경우)
+    private Integer id;                    // 댓글 ID
+    private Integer postId;                // 게시글 ID
+    private Integer parentId;              // 부모 댓글 ID (대댓글인 경우)
     private String content;             // 댓글 내용
     private String memberNickname;      // 작성자 닉네임
     private String rankName;            // 작성자 등급명
